+- add patches.arch/ppc-shmget02.patch +  cast size argument for shmget to signed (#152113)
suse-commit: 028e50d5b12ce0bc57b4937390a872682f42f2c6

+++ arch/powerpc/kernel/sys_ppc32.c
@@ -425,7 +425,7 @@ long compat_sys_ipc(u32 call, u32 first, u32 second, u32 third, compat_uptr_t pt
 		return sys_shmdt(compat_ptr(ptr));
 	case SHMGET:
 		/* sign extend key_t */
-		return sys_shmget((int)first, second, third);
+		return sys_shmget((int)first, (int)second, third);
 	case SHMCTL:
 		/* sign extend shmid */
 		return compat_sys_shmctl((int)first, second, compat_ptr(ptr));
